Renamed WatcherServices to LongRunningServices
